chore: purge orphaned modules
