--- src/keymap.rs
@@ -1,8 +1,0 @@
-pub fn default_keymap() -> &'static str {
-    r#"
-    Ctrl+K = Show Keymap
-    Ctrl+I = Toggle Triage Panel
-    Alt+Space / Alt+/ = Spotlight
-    Ctrl+M = Mode Switch
-    "#.trim()
-}
