+++ src/lib.rs
@@ -18,7 +18,6 @@ pub mod trust;
 pub mod federation;
 pub mod snapshot;
 pub mod retire;
-pub mod keymap;
 pub mod spotlight;
 pub mod theme;
 pub mod clipboard;
@@ -31,7 +30,6 @@ pub mod canvas;
 pub mod beam_color;
 pub mod ui;
 pub mod gemx;
-pub mod routineforge;
 pub mod triage;
 pub mod settings;
 pub mod hotkeys;
