+++ src/module_switcher/mod.rs
@@ -1,3 +1,2 @@
 pub mod input;
-pub mod control;
 pub mod ui;
