--- src/plugins/interface.rs
@@ -1,8 +1,0 @@
-use ratatui::{backend::CrosstermBackend, layout::Rect, Frame};
-use std::io::Stdout;
-
-pub type PluginFrame<'a> = Frame<'a, CrosstermBackend<Stdout>>;
-
-pub trait PluginRender {
-    fn render(&mut self, f: &mut PluginFrame<'_>, area: Rect);
-}
